add imageStore tests
