+++ weg-liTests/Store/ImagesStoreTests.swift
@@ -6,28 +6,54 @@
 //  Copyright © 2021 Martin Wilhelmi. All rights reserved.
 //
 
+@testable import weg_li
+import ComposableArchitecture
+import ComposableCoreLocation
 import XCTest
 
 class ImagesStoreTests: XCTestCase {
 
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
+    func test_addPhoto_shouldUpdateState() {
+            let store = TestStore(
+                initialState: ImagesViewState(
+                    showImagePicker: false,
+                    storedPhotos: [],
+                    resolvedLocation: nil
+                ),
+                reducer: imagesReducer,
+                environment: ImagesViewEnvironment(
+                    imageConverter: ImageConverterImplementation()
+                )
+            )
+            
+            let image = UIImage(systemName: "pencil")!
+            store.assert(
+                .send(.addPhoto(image)) {
+                    $0.storedPhotos = [
+                        StorableImage(uiImage: image)!
+                    ]
+                }
+            )
         }
-    }
-
+        
+        func test_removePhoto_shouldUpdateState() {
+            let image = UIImage(systemName: "pencil")!
+            let store = TestStore(
+                initialState: ImagesViewState(
+                    showImagePicker: false,
+                    storedPhotos: [StorableImage(uiImage: image)!],
+                    resolvedLocation: nil
+                ),
+                reducer: imagesReducer,
+                environment: ImagesViewEnvironment(
+                    imageConverter: ImageConverterImplementation()
+                )
+            )
+            
+            store.assert(
+                .send(.removePhoto(index: 0)) {
+                    $0.storedPhotos = []
+                }
+            )
+        }
 }
